feat: fix backend security, CORS, migrations, and frontend integration
Backend:
- Fix JWT filter double-registration by removing @Bean and creating filters inline in SecurityFilterChain
- Add CorsConfig.java for frontend origin (localhost:3000)
- Fix is_deleted NOT NULL constraint via AuditListener prePersist
- Add @DynamicInsert/@DynamicUpdate to BaseEntity
- Fix Admin/Audit controllers: change institutionId from @RequestParam to @RequestAttribute
- Fix AuditService compliance report: cast enum to name() instead of String
- Fix Flyway migration conflicts (V2, V3, V5, V30, V40)
- Remove context-path from application.yml

Frontend:
- Add certificateApi, adminApi, auditApi to lib/api.ts
- Add JWT decode and institutionId helpers to auth.tsx
- Rewrite certificate pages to use real API
- Add admin dashboard, settings, and roles pages
- Add audit page with tabs
- Fix TypeScript errors (educationLevel, live-classes count)
- Update sidebar with role-based nav for admin/audit

+++ backend/elmkusoma-core/src/main/resources/db/migration/V10__create_grading_tables.sql
@@ -8,8 +8,8 @@ CREATE TABLE grading_scales (
     name VARCHAR(100) NOT NULL,
     description VARCHAR(500),
     scale_type VARCHAR(50) NOT NULL, -- LETTER, NUMERIC, PERCENTAGE
-    min_value DECIMAL(5,2),
-    max_value DECIMAL(5,2),
+    min_value DOUBLE PRECISION,
+    max_value DOUBLE PRECISION,
     is_default BOOLEAN NOT NULL DEFAULT false,
     is_active BOOLEAN NOT NULL DEFAULT true,
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
